// src/app/shared/models/pks_timbangan.model.ts
export class PksTimbangan {
  id?: number;
  tipe?: string;
  no_tiket?: string;
  mill_id?: number;
  estate_id?: number;
  rayon_id?: number;
  divisi_id?: number;
  transportir_id?: number;
  item_id?: number;
  no_spat?: string;
  tanggal?: string;
  berat_bersih?: number;
  berat_kosong?: number;
  berat_potongan_persen?: number;
  berat_isi?: number;
  berat_potongan?: number;
  berat_terima?: number;
  jumlah_item?: number;
  jumlah_berondolan?: number;
  supplier_id?: number;
  no_plat?: string;
  nama_supir?: string;
  jam_masuk?: string;
  jam_keluar?: string;
  keterangan?: string;
  blok?: string;
  uoid?: string;



}